pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["HDYS-TTBYS <tatehide151@outlook.jp>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.0.6"
djangorestframework = "^3.12.2"
Pillow = "^8.1.0"
django-cors-headers = "^3.6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
